Clear the compiler warnings in ct002, and build the dashboard on Arduino too
A clean ESP32 build with every sub-block enabled and `-Wall -Wextra`
produced five warnings, all of them older than this branch:

  * three `%u` against a `uint32_t`, which is `long unsigned` on xtensa —
    cast at the call, lossless on both targets;
  * two `-Wdangling-reference` on `for (JsonObject d : root["data"]
    .as<JsonArray>())`. GCC 13 is wrong about the risk — the array is a
    handle into the document's own pool, not into the member proxy that
    dies at the end of the full expression — but the fix is to hold it in
    a named handle, which says so and keeps a strict build quiet.

ESPHome's own flags suppress several warning classes (-Wno-sign-compare,
-Wno-unused-variable and friends) and pass no -Wall of their own, so none
of these showed up in an ordinary build. Nothing added by this branch
warns at either level.

Also adds the dashboard on the Arduino framework to the compile matrix.
The schema accepts either ESP32 framework and web_server_base routes both
through web_server_idf, but only esp-idf was ever built — the component
keeps Arduino entries for its other sub-blocks for the same reason.

+++ esphome/components/ct002/ct002.cpp
@@ -775,7 +775,9 @@ void CT002Component::dump_config() {
   ESP_LOGCONFIG(TAG, "  CT MAC: %s", this->ct_mac_.empty() ? "(mirror)" : this->ct_mac_.c_str());
   ESP_LOGCONFIG(TAG, "  UDP Port: %u", this->udp_port_);
   ESP_LOGCONFIG(TAG, "  Active Control: %s", YESNO(this->active_control_));
-  ESP_LOGCONFIG(TAG, "  Max Sensor Age: %u ms", this->max_sensor_age_ms_);
+  // uint32_t is `long unsigned` on xtensa, so %u needs the cast to match
+  // (lossless — both are 32 bits here and on the host build).
+  ESP_LOGCONFIG(TAG, "  Max Sensor Age: %u ms", static_cast<unsigned>(this->max_sensor_age_ms_));
   for (uint8_t i = 0; i < this->num_phases_; ++i) {
     if (this->unit_scale_[i] != 1.0f) {
       ESP_LOGCONFIG(TAG, "  L%u Unit Scale: x%g (declared unit -> W)",
